+++ lib/ReactViews/DataCatalog/DataCatalogGroup.jsx
@@ -9,6 +9,10 @@ import getPath from "../../Core/getPath";
 import removeUserAddedData from "../../Models/Catalog/removeUserAddedData";
 import CatalogGroup from "./CatalogGroup";
 import DataCatalogMember from "./DataCatalogMember";
+import {
+  addRemoveButtonClicked,
+  allMappableMembersInWorkbench
+} from "./DisplayGroupHelper";
 
 const DataCatalogGroup = observer(
   createReactClass({
@@ -118,6 +122,20 @@ const DataCatalogGroup = observer(
             this.props.group
           )}
           selected={this.isSelected()}
+          // Pass these next three props down to deal with displayGroup functionality
+          displayGroup={group.displayGroup}
+          addRemoveButtonFunction={(event) => {
+            addRemoveButtonClicked(
+              this.props.group,
+              this.props.viewState,
+              this.props.terria,
+              event.shiftKey || event.ctrlKey
+            );
+          }}
+          allItemsLoaded={allMappableMembersInWorkbench(
+            this.props.group.members,
+            this.props.terria
+          )}
         >
           <If condition={this.isOpen()}>
             <For each="item" of={group.memberModels}>