+++ lib/ReactViews/Preview/GroupPreview.jsx
@@ -1,10 +1,7 @@
 import React from "react";
-
 import createReactClass from "create-react-class";
-
 import PropTypes from "prop-types";
 import { observer } from "mobx-react";
-
 import DataPreviewSections from "./DataPreviewSections";
 import DataPreviewUrl from "./DataPreviewUrl";
 import measureElement from "../HOCs/measureElement";
@@ -13,6 +10,10 @@ import parseCustomMarkdownToReact from "../Custom/parseCustomMarkdownToReact";
 import SharePanel from "../Map/Panels/SharePanel/SharePanel";
 import { withTranslation } from "react-i18next";
 import WarningBox from "./WarningBox";
+import {
+  addRemoveButtonClicked,
+  allMappableMembersInWorkbench
+} from "../DataCatalog/DisplayGroupHelper";
 
 /**
  * A "preview" for CatalogGroup.
@@ -44,7 +45,30 @@ const GroupPreview = observer(
             ref={(component) => (this.refToMeasure = component)}
           >
             <h3>{this.props.previewed.name}</h3>
+
             <div className={Styles.shareLinkWrapper}>
+              {/* If this is a display group, show the "Add/Remove All" button next to the shareLink */}
+              {this.props.previewed.displayGroup === true && (
+                <button
+                  type="button"
+                  onClick={(event) => {
+                    addRemoveButtonClicked(
+                      this.props.previewed,
+                      this.props.viewState,
+                      this.props.terria,
+                      event.shiftKey || event.ctrlKey
+                    );
+                  }}
+                  className={Styles.btnAddAll}
+                >
+                  {allMappableMembersInWorkbench(
+                    this.props.previewed.members,
+                    this.props.terria
+                  )
+                    ? t("models.catalog.removeAll")
+                    : t("models.catalog.addAll")}
+                </button>
+              )}
               <SharePanel
                 catalogShare
                 modalWidth={this.props.widthFromMeasureElementHOC}